+++ composeApp/src/commonMain/kotlin/weather/data/dto/neighbor_weather/Hourly.kt
@@ -11,7 +11,7 @@ data class Hourly(
     @SerialName("temperature_2m")
     val temperature2m: List<Double>,
     @SerialName("relative_humidity_2m")
-    val relativeHumidity2m: List<Int>,
+    val relativeHumidity2m: List<Double>,
     @SerialName("apparent_temperature")
     val apparentTemperature: List<Double>,
     @SerialName("precipitation")
