+++ composeApp/src/commonMain/kotlin/weather/data/dto/neighbor_weather/NeighborWeatherDto.kt
@@ -19,7 +19,7 @@ data class NeighborWeatherDto(
     @SerialName("timezone_abbreviation")
     val timezoneAbbreviation: String,
     @SerialName("elevation")
-    val elevation: Int,
+    val elevation: Double,
     @SerialName("current_units")
     val currentUnits: CurrentUnits,
     @SerialName("current")
