+++ composeApp/src/commonMain/kotlin/weather/domain/model/CurrentWeather.kt
@@ -5,7 +5,7 @@ import kotlinx.datetime.LocalDateTime
 data class CurrentWeather(
     val time: LocalDateTime,
     val temperature: Double,
-    val relativeHumidity: Int,
+    val relativeHumidity: Double,
     val apparentTemperature: Double,
     val precipitation: Double,
     val weatherCode: Int,
