+++ composeApp/src/commonMain/kotlin/weather/domain/model/HourlyWeather.kt
@@ -5,7 +5,7 @@ import kotlinx.datetime.LocalDateTime
 data class HourlyWeather(
     val time: List<LocalDateTime>,
     val temperature: List<Double>,
-    val relativeHumidity: List<Int>,
+    val relativeHumidity: List<Double>,
     val apparentTemperature: List<Double>,
     val precipitation: List<Double>,
     val weatherCode: List<Int>
